demo/stable-diffusion/requirements.txt
accelerate==1.3.0
aiofiles==23.2.1
annotated-types==0.7.0
anyio==4.8.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
diffusers==0.32.2
exceptiongroup==1.2.2
fastapi==0.115.7
ffmpy==0.5.0
filelock==3.17.0
fsspec==2024.12.0
ftfy==6.3.1
gradio==5.13.1
gradio_client==1.6.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
idna==3.10
importlib_metadata==8.6.1
Jinja2==3.1.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.2
nvidia-ml-py3==7.352.0
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pillow==11.1.0
psutil==6.1.1
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
ruff==0.9.3
safehttpx==0.1.6
safetensors==0.5.2
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.45.2
sympy==1.13.1
tokenizers==0.21.0
tomlkit==0.13.2
torch==2.5.1
tqdm==4.67.1
transformers==4.48.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
wcwidth==0.2.13
websockets==14.2
zipp==3.21.0
